Updates related to GPIO driver

+++ COTS/HAL/LED/LED_Cfg.h
@@ -20,7 +20,7 @@
 
 typedef struct
 {
-    u8 gpioPort;
+    u32 gpioPort;
     u8 pin;
     u8 mode;
     u8 speed;
